+++ boards/Kconfig
@@ -681,6 +681,14 @@ config ARCH_BOARD_ICICLE_MPFS
                 This is the board configuration for the port of NuttX to the
                 MicroChip icicle-mpfs board. This board features the RISC-V MPFS.
 
+config ARCH_BOARD_M100PFSEVP_MPFS
+        bool "Aries M100PFSEVP evaluation platform for MPFS"
+        depends on ARCH_CHIP_MPFS
+        select CONFIG_MPFS_EMMCSD_MUX_GPIO if !MPFS_WITH_QEMU
+        ---help---
+                This is the board configuration for the port of NuttX to the
+                MicroChip m100pfsevp-mpfs board. This board features the RISC-V MPFS.
+
 config ARCH_BOARD_MAX32660_EVSYS
 	bool "Maxim Integrated MAX32660-EVSYS"
 	depends on ARCH_CHIP_MAX32660
@@ -2389,6 +2397,7 @@ config ARCH_BOARD
 	default "maix-bit"                 if ARCH_BOARD_MAIX_BIT
 	default "smartl-c906"              if ARCH_BOARD_SMARTL_C906
 	default "icicle"                   if ARCH_BOARD_ICICLE_MPFS
+	default "m100pfsevp"               if ARCH_BOARD_M100PFSEVP_MPFS
 	default "maple"                    if ARCH_BOARD_MAPLE
 	default "makerlisp"                if ARCH_BOARD_MAKERLISP
 	default "max32660-evsys"           if ARCH_BOARD_MAX32660_EVSYS
@@ -3198,6 +3207,9 @@ source "boards/risc-v/c906/smartl-c906/Kconfig"
 endif
 if ARCH_BOARD_ICICLE_MPFS
 source "boards/risc-v/mpfs/icicle/Kconfig"
+endif
+if ARCH_BOARD_M100PFSEVP_MPFS
+source "boards/risc-v/mpfs/m100pfsevp/Kconfig"
 endif
 if ARCH_BOARD_RV32M1_VEGA
 source "boards/risc-v/rv32m1/rv32m1-vega/Kconfig"
