+++ boards/risc-v/mpfs/common/kernel/Makefile
@@ -1,5 +1,5 @@
 ############################################################################
-# boards/risc-v/mpfs/icicle//kernel/Makefile
+# boards/risc-v/mpfs/common/kernel/Makefile
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
