+++ boards/risc-v/mpfs/common/src/Make.defs
@@ -1,0 +1,45 @@
+############################################################################
+# boards/risc-v/mpfs/common/src/Make.defs
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.  The
+# ASF licenses this file to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance with the
+# License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+############################################################################
+
+CSRCS += mpfs_boot.c
+
+ifeq ($(CONFIG_I2C),y)
+CSRCS += mpfs_i2c.c
+endif
+
+ifeq ($(CONFIG_ARCH_FPU),y)
+CSRCS += mpfs_ostest.c
+endif
+
+ifeq ($(CONFIG_SPI),y)
+CSRCS += mpfs_board_spi.c
+endif
+
+ifeq ($(CONFIG_MPFS_HAVE_COREPWM),y)
+CSRCS += mpfs_pwm.c
+endif
+
+ifeq ($(CONFIG_MPFS_EMMCSD),y)
+CSRCS += mpfs_emmcsd.c
+endif
+
+DEPPATH += --dep-path src
+VPATH += :src
+CFLAGS += $(shell $(INCDIR) "$(CC)" $(TOPDIR)$(DELIM)arch$(DELIM)$(CONFIG_ARCH)$(DELIM)src$(DELIM)board$(DELIM)src)
