+++ boards/risc-v/mpfs/common/src/mpfs_pwm.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/risc-v/mpfs/icicle/src/mpfs_pwm.c
+ * boards/risc-v/mpfs/common/src/mpfs_pwm.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
