+++ boards/risc-v/mpfs/icicle/src/board_config.h
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/risc-v/mpfs/icicle/src/mpfsicicle.h
+ * boards/risc-v/mpfs/icicle/src/board_config.h
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -18,8 +18,8 @@
  *
  ****************************************************************************/
 
-#ifndef __BOARDS_RISCV_ICICLE_MPFS_SRC_MPFSICICLE_H
-#define __BOARDS_RISCV_ICICLE_MPFS_SRC_MPFSICICLE_H
+#ifndef __BOARDS_RISCV_ICICLE_MPFS_SRC_BOARD_CONFIG_H
+#define __BOARDS_RISCV_ICICLE_MPFS_SRC_BOARD_CONFIG_H
 
 /****************************************************************************
  * Included Files
@@ -47,4 +47,4 @@ int mpfs_board_i2c_init(void);
 int mpfs_board_emmcsd_init(void);
 int mpfs_pwm_setup(void);
 
-#endif /* __BOARDS_RISCV_ICICLE_MPFS_SRC_MPFSICICLE_H */
+#endif /* __BOARDS_RISCV_ICICLE_MPFS_SRC_BOARD_CONFIG_H */
