+++ boards/risc-v/mpfs/m100pfsevp/Kconfig
@@ -1,0 +1,8 @@
+#
+# For a description of the syntax of this configuration file,
+# see misc/tools/kconfig-language.txt.
+#
+
+if ARCH_BOARD_M100PFSEVP_MPFS
+
+endif
