+++ boards/risc-v/mpfs/m100pfsevp/src/mpfs_bringup.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/risc-v/mpfs/icicle/src/mpfs_i2c.c
+ * boards/risc-v/mpfs/m100pfsevp/src/mpfs_bringup.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -24,78 +24,81 @@
 
 #include <nuttx/config.h>
 
+#include <sys/mount.h>
+#include <stdbool.h>
+#include <stdio.h>
 #include <debug.h>
 #include <errno.h>
-#include <sys/types.h>
-#include <nuttx/i2c/i2c_master.h>
 
-#include "mpfs_i2c.h"
+#include <nuttx/board.h>
+#include <nuttx/drivers/ramdisk.h>
+
+#include "board_config.h"
+#include "mpfs_corepwm.h"
+#include "mpfs.h"
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: board_i2c_init
- *
- * Description:
- *   Configure the I2C driver.
- *
- * Returned Value:
- *   Zero (OK) is returned on success; A negated errno value is returned
- *   to indicate the nature of any failure.
- *
+ * Name: mpfs_bringup
  ****************************************************************************/
 
-int mpfs_board_i2c_init(void)
+int mpfs_bringup(void)
 {
   int ret = OK;
 
-#if defined(CONFIG_MPFS_I2C0) || defined(CONFIG_MPFS_I2C1)
-#ifdef CONFIG_I2C_DRIVER
-  int bus = 0;
-#endif
-  FAR struct i2c_master_s *i2c;
-#endif
+#if defined(CONFIG_I2C_DRIVER)
+  /* Configure I2C peripheral interfaces */
 
-#ifdef CONFIG_MPFS_I2C0
-  i2c = mpfs_i2cbus_initialize(0);
+  ret = mpfs_board_i2c_init();
 
-  if (i2c == NULL)
-    {
-      i2cerr("ERROR: Failed to init I2C0 interface\n");
-      return -ENODEV;
-    }
-
-#ifdef CONFIG_I2C_DRIVER
-  ret = i2c_register(i2c, bus++);
   if (ret < 0)
     {
-      i2cerr("ERROR: Failed to register I2C0 driver: %d\n", ret);
-      mpfs_i2cbus_uninitialize(i2c);
-      return ret;
+      syslog(LOG_ERR, "Failed to initialize I2C driver: %d\n", ret);
     }
 #endif
-#endif
 
-#ifdef CONFIG_MPFS_I2C1
-  i2c = mpfs_i2cbus_initialize(1);
+#ifdef CONFIG_FS_PROCFS
+  /* Mount the procfs file system */
 
-  if (i2c == NULL)
-    {
-      i2cerr("ERROR: Failed to init I2C1 interface\n");
-      return -ENODEV;
-    }
-
-#ifdef CONFIG_I2C_DRIVER
-  ret = i2c_register(i2c, bus);
+  ret = mount(NULL, "/proc", "procfs", 0, NULL);
   if (ret < 0)
     {
-      i2cerr("ERROR: Failed to register I2C1 driver: %d\n", ret);
-      mpfs_i2cbus_uninitialize(i2c);
-      return ret;
+      serr("ERROR: Failed to mount procfs at %s: %d\n", "/proc", ret);
     }
 #endif
+
+#if defined(CONFIG_MPFS_SPI0) || defined(CONFIG_MPFS_SPI1)
+  /* Configure SPI peripheral interfaces */
+
+  ret = mpfs_board_spi_init();
+
+  if (ret < 0)
+    {
+      syslog(LOG_ERR, "Failed to initialize SPI driver: %d\n", ret);
+    }
+#endif
+
+#ifdef CONFIG_MPFS_HAVE_COREPWM
+  /* Configure PWM peripheral interfaces */
+
+  ret = mpfs_pwm_setup();
+
+  if (ret < 0)
+    {
+      syslog(LOG_ERR, "Failed to initialize CorePWM driver: %d\n", ret);
+    }
+#endif
+
+#ifdef CONFIG_MPFS_EMMCSD
+  ret = mpfs_board_emmcsd_init();
+
+  if (ret < 0)
+    {
+      syslog(LOG_ERR, "Failed to init eMMCSD driver: %d\n", ret);
+    }
 #endif
 
   return ret;
